+++ R/crosstable.R
@@ -24,6 +24,7 @@ crosstable_caller = rlang::env()
 #' @param showNA Whether to show NA in categorical variables (one of \code{c("ifany", "always", "no")}, like in \code{table()}).
 #' @param label Whether to show labels. See [import_labels()] or [set_label()]for how to add labels to the dataset columns.
 #' @param cor_method One of `c("pearson", "kendall", "spearman")` to indicate which correlation coefficient is to be used.
+#' @param drop_levels Whether to drop unused levels of factor variables. Default to `TRUE`.
 #' @param times When using formula with [survival::Surv()] objects, which times to summarize.
 #' @param followup When using formula with [survival::Surv()] objects, whether to display follow-up time.
 #' @param test Whether to perform tests.
@@ -104,6 +105,7 @@ crosstable = function(data, cols=everything(), ..., by=NULL,
                       showNA = c("ifany", "always", "no"), label = TRUE,
                       funs = c(" " = cross_summary), funs_arg=list(),
                       cor_method = c("pearson", "kendall", "spearman"),
+                      drop_levels = TRUE,
                       unique_numeric = 3, date_format=NULL,
                       times = NULL, followup = FALSE,
                       test = FALSE, test_args = crosstable_test_args(),
@@ -422,7 +424,7 @@ crosstable = function(data, cols=everything(), ..., by=NULL,
 
     rtn = cross_by(data_x=data_x, data_y=data_y2, funs=funs, funs_arg=funs_arg,
                    percent_pattern=percent_pattern, total=total,
-                   percent_digits=percent_digits, showNA=showNA,
+                   percent_digits=percent_digits, showNA=showNA, drop_levels=drop_levels,
                    cor_method=cor_method, times=times, followup=followup, test=test, test_args=test_args,
                    effect=effect, effect_args=effect_args, label=label)
 
@@ -432,7 +434,7 @@ crosstable = function(data, cols=everything(), ..., by=NULL,
     data_y2 = data_y
     rtn = cross_by(data_x=data_x, data_y=data_y, funs=funs, funs_arg=funs_arg,
                    percent_pattern=percent_pattern, percent_digits=percent_digits,
-                   total=total, showNA=showNA,
+                   total=total, showNA=showNA, drop_levels=drop_levels,
                    cor_method=cor_method, times=times, followup=followup, test=test, test_args=test_args,
                    effect=effect, effect_args=effect_args, label=label)
     class(rtn) = c("crosstable", "tbl_df", "tbl", "data.frame")