+++ NEWS.md
@@ -19,6 +19,9 @@ Crosstables for descriptive analyses. See documentation at <https://danchaltiel.
     ```
 -   New function `body_add_table_list()` that replaces `body_add_crosstable_list()` and `body_add_flextable_list()`. It allows crosstables and flextable (and even dataframes, which are turned to flextables first) and has a new argument `fun_after` to control what comes after the table. See examples in `?body_add_table_list`.
 
+#### Breaking changes (minor)
+
+-   New argument `crosstable(drop_levels)` to drop unused levels in factors. Default to `TRUE` even though it is new.
 
 #### Bug fixes